Use `*WithAssertError` convention if the operator throws an error as assertion for unexpected type.
This sort the our convention with the changing in #2307.

Exist operators that have `*WithEnsureType` style naming have 2 manner:

1. Just Ensure types by a type predicate.
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/maybe/filter.ts#L31
     - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/plain_result/is_ok_and.ts#L26
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/plain_result/is_err_and.ts#L26
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/undefinable/filter.ts#L34
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/nullable/filter.ts#L31
1. Throw an error as like as [assertion](https://en.wikipedia.org/wiki/Assertion_(software_development)) if they encounters an unexpected type.
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/plain_result/try_catch.ts#L40
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/plain_result/try_catch_async.ts#L45
    - https://github.com/option-t/option-t/blob/97164733bfa222727eb53a61dfe472f7c1cf52f3/packages/option-t/src/plain_result/deprecated/unwrap_or_throw_error.ts#L39

This mixing confuse user (including me).
The laters should contains "assert" as their naming.
